requirements.txt
google-generativeai
pandas
streamlit
python-dotenv